Done with day 1 of AoC-2018
The first problem was to add toggether 972 numbers and.
The second problem was to find the first number that repeted when adding the numbers.
First i made a function to read the text of a file becuase it was to much text to put in an array.
Then i made the string in to a string array and then into
an int array becuase i needed to seperate the numbers so i could add them toggether.
Then for the second part of the problem i made a loop
that would see if the sum of the curent number matched any number that had come before.